sketch of a session implementation
The `session` field will only be available when authenticated via
keycloak. In order to use the `userBySub` relation, a `User` must
already exist with the `user.uuid` column set to the `session.sub`.

example usage:
```graphql
{
  session {
    jti
    exp
    nbf
    iat
    iss
    aud
    sub
    typ
    azp
    authTime
    sessionState
    acr
    emailVerified
    name
    preferredUsername
    givenName
    familyName
    email
    userBySub {
      id
      rowId
      uuid
      firstName
      lastName
      emailAddress
      createdAt
      updatedAt
      createdBy
      updatedBy
      userOrganisationsByUserId {
        edges {
          node {
            id
          }
        }
      }
    }
  }
}
```

Comments describing the `JwtToken` type conformance to [RFC 7519] are
still needed, and the nullability of those fields should be correctly
set in the type to conform to the RFC.

[RFC 7519]: https://tools.ietf.org/html/rfc7519