overriding resolution of "open" should happen in yarn.lock
@hamzajaved bump is happening because [you pinned `open` in `package.json`](https://github.com/bcgov/cas-ciip-portal/commit/0f2ec178ec063d9858c5cce9f2c352c183c7ffdd)
after a security vulnerability warning.

However, it's not actually required in our `devDependenies`, it's
required by dependencies of our devDependencies. As a result, we
actually want to correct this vulnerability warning by using
[yarn resolutions](https://yarnpkg.com/lang/en/docs/selective-version-resolutions/)
so the transitive dependency itself is updated and then send a PR to the
upstream project so the fix is committed for all.