feature(session): integrate JwtToken session into default layout
A the root-level 'session' field is now integrated into the Header
component seen on most pages. This allows for login and logout from a
consistent location on all screens.

Bug: The session field is currently only rendering on client-side relay
requests since the session cookie isn't available to the server-side
relay renderer for some reason.

This touches a whole lot of linter errors since pre-commit hooks are now
installed.