fix(security): patch arbitrary code execution exploit
Versions of handlebars prior to 4.5.3 are vulnerable to Arbitrary Code
Execution. The package's lookup helper fails to properly validate
templates, allowing attackers to submit templates that execute arbitrary
JavaScript in the system. It is due to an incomplete fix for a previous
issue. This vulnerability can be used to run arbitrary code in a server
processing Handlebars templates or on a victim's browser
(effectively serving as Cross-Site Scripting).

see https://www.npmjs.com/advisories/1324
see https://www.npmjs.com/advisories/1316